+++ RELEASE_NOTES.md
@@ -1,5 +1,22 @@
 # SharpSCCM Release Notes
 
+### Version 2.0.12 (September 16, 2024)
+##### Changes
+- Fixed case matching issue in exec command
+
+### Version 2.0.11 (July 3, 2024)
+##### Changes
+- Added working directory to exec command
+- Added -mp back to invoke command
+
+### Version 2.0.10 (April 15, 2024)
+##### Changes
+- Added get admins command to list SCCM admin users
+
+### Version 2.0.9 (April 15, 2024)
+##### Changes
+- Added option to deobfuscate a secret string offline
+
 ### Version 2.0.8 (March 19, 2024)
 ##### Changes
 - Fix from @subat0mik for machines that require FIPS-compliant algorithms (PR #53)